requirements.txt
attrs>=19.2.0
pytest==5.4.3
lxml==4.5.2  # For mypy HTML reporting
mypy==0.782
hypothesis==5.16.1
pytest-cov==2.10.0
